functions/notes_api/requirements.txt
functions-framework==3.*
firebase-admin==6.*
google-cloud-firestore==2.*
google-cloud-storage==2.*
flask==3.*